refactor: move the feature flag conditional to the parent component
