+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/layout/dot-edit-layout/dot-edit-layout.component.spec.ts
@@ -15,6 +15,7 @@ import { DotTemplateContainersCacheService } from '@dotcms/app/api/services/dot-
 import {
     DotMessageService,
     DotPageLayoutService,
+    DotPropertiesService,
     DotSessionStorageService
 } from '@dotcms/data-access';
 import { DotCMSResponse, HttpCode, ResponseView } from '@dotcms/dotcms-js';
@@ -67,6 +68,7 @@ describe('DotEditLayoutComponent', () => {
     let dotHttpErrorManagerService: DotHttpErrorManagerService;
     let dotEditLayoutService: DotEditLayoutService;
     let dotSessionStorageService: DotSessionStorageService;
+    let dotPropertiesService: DotPropertiesService;
     let router: Router;
 
     beforeEach(waitForAsync(() => {
@@ -125,6 +127,12 @@ describe('DotEditLayoutComponent', () => {
                             }
                         }
                     }
+                },
+                {
+                    provide: DotPropertiesService,
+                    useValue: {
+                        getKey: () => of('false')
+                    }
                 }
             ]
         });
@@ -133,8 +141,6 @@ describe('DotEditLayoutComponent', () => {
     beforeEach(() => {
         fixture = TestBed.createComponent(DotEditLayoutComponent);
         component = fixture.componentInstance;
-        layoutDesignerDe = fixture.debugElement.query(By.css('dot-edit-layout-designer'));
-        layoutDesigner = layoutDesignerDe.componentInstance;
 
         dotPageLayoutService = TestBed.inject(DotPageLayoutService);
         dotGlobalMessageService = TestBed.inject(DotGlobalMessageService);
@@ -142,128 +148,154 @@ describe('DotEditLayoutComponent', () => {
         dotHttpErrorManagerService = TestBed.inject(DotHttpErrorManagerService);
         dotEditLayoutService = TestBed.inject(DotEditLayoutService);
         dotSessionStorageService = TestBed.inject(DotSessionStorageService);
+        dotPropertiesService = TestBed.inject(DotPropertiesService);
         router = TestBed.inject(Router);
-
-        fakeLayout = {
-            body: null,
-            footer: false,
-            header: true,
-            sidebar: null,
-            themeId: '123',
-            title: 'Title',
-            width: '100'
-        };
-
-        fixture.detectChanges();
     });
 
-    it('should be 100% min-width in the host', () => {
-        // https://github.com/dotCMS/core/issues/19540
-        expect(fixture.debugElement.nativeElement.style.minWidth).toBe('100%');
-    });
-
-    it('should pass attr to the dot-edit-layout-designer', () => {
-        const state = mockDotRenderedPage();
-        expect(layoutDesigner.layout).toEqual(state.layout);
-        expect(layoutDesigner.theme).toEqual(state.template.theme);
-        expect(layoutDesigner.title).toEqual(state.page.title);
-        expect(layoutDesigner.apiLink).toEqual('api/v1/page/render/an/url/test?language_id=1');
-        expect(layoutDesigner.url).toEqual(state.page.pageURI);
-    });
-
-    describe('save', () => {
-        it('should save the layout', () => {
-            const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
-            spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
-
-            layoutDesignerDe.triggerEventHandler('save', fakeLayout);
-
-            expect(dotGlobalMessageService.loading).toHaveBeenCalledWith('Saving');
-            expect(dotGlobalMessageService.success).toHaveBeenCalledWith('Saved');
-            expect(dotGlobalMessageService.error).not.toHaveBeenCalled();
-
-            expect(dotPageLayoutService.save).toHaveBeenCalledWith('123', {
-                ...fakeLayout,
-                title: null
-            });
-            expect(dotTemplateContainersCacheService.set).toHaveBeenCalledWith({
-                '/default/': processedContainers[0].container,
-                '/banner/': processedContainers[1].container
-            });
-            expect(component.pageState).toEqual(new DotPageRender(mockDotRenderedPage()));
+    describe('with data', () => {
+        beforeEach(() => {
+            fixture.detectChanges();
+            layoutDesignerDe = fixture.debugElement.query(By.css('dot-edit-layout-designer'));
+            layoutDesigner = layoutDesignerDe.componentInstance;
+            fakeLayout = {
+                body: null,
+                footer: false,
+                header: true,
+                sidebar: null,
+                themeId: '123',
+                title: 'Title',
+                width: '100'
+            };
         });
 
-        it('should save the layout after 10000', fakeAsync(() => {
-            const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
-            spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
+        it('should be 100% min-width in the host', () => {
+            // https://github.com/dotCMS/core/issues/19540
+            expect(fixture.debugElement.nativeElement.style.minWidth).toBe('100%');
+        });
 
-            layoutDesignerDe.triggerEventHandler('updateTemplate', fakeLayout);
+        it('should pass attr to the dot-edit-layout-designer', () => {
+            fixture.detectChanges();
+            const state = mockDotRenderedPage();
+            expect(layoutDesigner.layout).toEqual(state.layout);
+            expect(layoutDesigner.theme).toEqual(state.template.theme);
+            expect(layoutDesigner.title).toEqual(state.page.title);
+            expect(layoutDesigner.apiLink).toEqual('api/v1/page/render/an/url/test?language_id=1');
+            expect(layoutDesigner.url).toEqual(state.page.pageURI);
+        });
 
-            tick(10000);
-            expect(dotGlobalMessageService.loading).toHaveBeenCalledWith('Saving');
-            expect(dotGlobalMessageService.success).toHaveBeenCalledWith('Saved');
-            expect(dotGlobalMessageService.error).not.toHaveBeenCalled();
-
-            expect(dotPageLayoutService.save).toHaveBeenCalledWith('123', {
-                ...fakeLayout,
-                title: null
+        describe('save', () => {
+            beforeEach(() => {
+                fixture.detectChanges();
             });
-            expect(dotTemplateContainersCacheService.set).toHaveBeenCalledWith({
-                '/default/': processedContainers[0].container,
-                '/banner/': processedContainers[1].container
+
+            it('should save the layout', () => {
+                const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
+                spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
+
+                layoutDesignerDe.triggerEventHandler('save', fakeLayout);
+
+                expect(dotGlobalMessageService.loading).toHaveBeenCalledWith('Saving');
+                expect(dotGlobalMessageService.success).toHaveBeenCalledWith('Saved');
+                expect(dotGlobalMessageService.error).not.toHaveBeenCalled();
+
+                expect(dotPageLayoutService.save).toHaveBeenCalledWith('123', {
+                    ...fakeLayout,
+                    title: null
+                });
+                expect(dotTemplateContainersCacheService.set).toHaveBeenCalledWith({
+                    '/default/': processedContainers[0].container,
+                    '/banner/': processedContainers[1].container
+                });
+                expect(component.pageState).toEqual(new DotPageRender(mockDotRenderedPage()));
             });
-            expect(component.pageState).toEqual(new DotPageRender(mockDotRenderedPage()));
-        }));
 
-        it('should not save the layout when observable is destroy', fakeAsync(() => {
-            const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
-            spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
+            it('should save the layout after 10000', fakeAsync(() => {
+                const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
+                spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
 
-            // Destroy should be true.
-            component.destroy$.subscribe((value) => expect(value).toBeTruthy());
+                layoutDesignerDe.triggerEventHandler('updateTemplate', fakeLayout);
 
-            // Trigger the observable
-            layoutDesignerDe.triggerEventHandler('updateTemplate', fakeLayout);
+                tick(10000);
+                expect(dotGlobalMessageService.loading).toHaveBeenCalledWith('Saving');
+                expect(dotGlobalMessageService.success).toHaveBeenCalledWith('Saved');
+                expect(dotGlobalMessageService.error).not.toHaveBeenCalled();
 
-            // Destroy the observable
-            component.destroy$.next(true);
-            component.destroy$.complete();
-            tick(10000);
+                expect(dotPageLayoutService.save).toHaveBeenCalledWith('123', {
+                    ...fakeLayout,
+                    title: null
+                });
+                expect(dotTemplateContainersCacheService.set).toHaveBeenCalledWith({
+                    '/default/': processedContainers[0].container,
+                    '/banner/': processedContainers[1].container
+                });
+                expect(component.pageState).toEqual(new DotPageRender(mockDotRenderedPage()));
+            }));
 
-            expect(dotPageLayoutService.save).not.toHaveBeenCalled();
-        }));
+            it('should not save the layout when observable is destroy', fakeAsync(() => {
+                const res: DotPageRender = new DotPageRender(mockDotRenderedPage());
+                spyOn(dotPageLayoutService, 'save').and.returnValue(of(res));
 
-        it('should handle error when save fail', (done) => {
-            spyOn(dotPageLayoutService, 'save').and.returnValue(
-                throwError(
-                    new ResponseView(
-                        new HttpResponse<DotCMSResponse<unknown>>(
-                            mockResponseView(HttpCode.BAD_REQUEST)
+                // Destroy should be true.
+                component.destroy$.subscribe((value) => expect(value).toBeTruthy());
+
+                // Trigger the observable
+                layoutDesignerDe.triggerEventHandler('updateTemplate', fakeLayout);
+
+                // Destroy the observable
+                component.destroy$.next(true);
+                component.destroy$.complete();
+                tick(10000);
+
+                expect(dotPageLayoutService.save).not.toHaveBeenCalled();
+            }));
+
+            it('should handle error when save fail', (done) => {
+                spyOn(dotPageLayoutService, 'save').and.returnValue(
+                    throwError(
+                        new ResponseView(
+                            new HttpResponse<DotCMSResponse<unknown>>(
+                                mockResponseView(HttpCode.BAD_REQUEST)
+                            )
                         )
                     )
-                )
-            );
+                );
 
-            layoutDesignerDe.triggerEventHandler('save', fakeLayout);
-            expect(dotGlobalMessageService.error).toHaveBeenCalledWith('Unknown Error');
-            expect(dotHttpErrorManagerService.handle).toHaveBeenCalledTimes(1);
-            dotEditLayoutService.canBeDesactivated$.subscribe((resp) => {
-                expect(resp).toBeTruthy();
-                done();
+                layoutDesignerDe.triggerEventHandler('save', fakeLayout);
+                expect(dotGlobalMessageService.error).toHaveBeenCalledWith('Unknown Error');
+                expect(dotHttpErrorManagerService.handle).toHaveBeenCalledTimes(1);
+                dotEditLayoutService.canBeDesactivated$.subscribe((resp) => {
+                    expect(resp).toBeTruthy();
+                    done();
+                });
+            });
+
+            it('should remove variant key from session storage on destoy', () => {
+                spyOn(dotSessionStorageService, 'removeVariantId');
+                component.ngOnDestroy();
+                expect(dotSessionStorageService.removeVariantId).toHaveBeenCalledTimes(1);
+            });
+
+            it('should keep variant key from session storage if going to Edit content portlet', () => {
+                router.routerState.snapshot.url = '/edit-page/content';
+                spyOn(dotSessionStorageService, 'removeVariantId');
+                component.ngOnDestroy();
+                expect(dotSessionStorageService.removeVariantId).toHaveBeenCalledTimes(0);
             });
         });
+    });
 
-        it('should remove variant key from session storage on destoy', () => {
-            spyOn(dotSessionStorageService, 'removeVariantId');
-            component.ngOnDestroy();
-            expect(dotSessionStorageService.removeVariantId).toHaveBeenCalledTimes(1);
+    describe('New Template Builder', () => {
+        beforeEach(() => {
+            spyOn(dotPropertiesService, 'getKey').and.returnValue(of('true'));
         });
 
-        it('should keep variant key from session storage if going to Edit content portlet', () => {
-            router.routerState.snapshot.url = '/edit-page/content';
-            spyOn(dotSessionStorageService, 'removeVariantId');
-            component.ngOnDestroy();
-            expect(dotSessionStorageService.removeVariantId).toHaveBeenCalledTimes(0);
+        it('should show new template builder component', () => {
+            fixture.detectChanges();
+            const component: DebugElement = fixture.debugElement.query(
+                By.css('[data-testId="new-template-builder"]')
+            );
+
+            expect(component).toBeTruthy();
         });
     });
 });
