+++ core-web/apps/dotcms-ui/src/app/portlets/dot-templates/dot-template-create-edit/dot-template-builder/dot-template-builder.component.html
@@ -1,7 +1,7 @@
 <p-tabView>
     <p-tabPanel
-        data-testId="builder"
         [header]="item.type === 'advanced' ? ('code' | dm) : ('design' | dm)"
+        data-testId="builder"
     >
         <ng-container *ngIf="item.type === 'advanced'; else elseBlock">
             <dot-template-advanced
@@ -13,20 +13,26 @@
             ></dot-template-advanced>
         </ng-container>
         <ng-template #elseBlock>
-            <dot-edit-layout-designer
-                [theme]="item.theme"
-                [disablePublish]="item.live"
-                [layout]="item.layout"
-                (saveAndPublish)="saveAndPublish.emit($event)"
-                (updateTemplate)="updateTemplate.emit($event)"
-                (save)="save.emit($event)"
-            ></dot-edit-layout-designer>
+            <ng-container *ngIf="enableNewBuilder$ | async as context">
+                <h1 *ngIf="context.enableNewBuilder" data-testId="new-template-builder">
+                    New Template Builder
+                </h1>
+                <dot-edit-layout-designer
+                    *ngIf="!context.enableNewBuilder"
+                    [theme]="item.theme"
+                    [disablePublish]="item.live"
+                    [layout]="item.layout"
+                    (saveAndPublish)="saveAndPublish.emit($event)"
+                    (updateTemplate)="updateTemplate.emit($event)"
+                    (save)="save.emit($event)"
+                ></dot-edit-layout-designer>
+            </ng-container>
         </ng-template>
     </p-tabPanel>
     <p-tabPanel [header]="'Permissions' | dm">
         <ng-template pTemplate="content">
             <dot-portlet-box>
-                <dot-iframe data-testId="permissionsIframe" [src]="permissionsUrl"></dot-iframe>
+                <dot-iframe [src]="permissionsUrl" data-testId="permissionsIframe"></dot-iframe>
             </dot-portlet-box>
         </ng-template>
     </p-tabPanel>
@@ -35,9 +41,9 @@
             <dot-portlet-box>
                 <dot-iframe
                     #historyIframe
+                    [src]="historyUrl"
                     (custom)="custom.emit($event)"
                     data-testId="historyIframe"
-                    [src]="historyUrl"
                 ></dot-iframe>
             </dot-portlet-box>
         </ng-template>
