+++ core-web/apps/dotcms-ui/src/app/portlets/dot-templates/dot-template-create-edit/dot-template-builder/dot-template-builder.component.ts
@@ -1,3 +1,5 @@
+import { Observable } from 'rxjs';
+
 import {
     Component,
     EventEmitter,
@@ -8,7 +10,11 @@ import {
     ViewChild
 } from '@angular/core';
 
+import { map } from 'rxjs/operators';
+
 import { IframeComponent } from '@components/_common/iframe/iframe-component';
+import { DotPropertiesService } from '@dotcms/data-access';
+import { FeaturedFlags } from '@dotcms/dotcms-models';
 
 import { DotTemplateItem } from '../store/dot-template.store';
 
@@ -29,7 +35,14 @@ export class DotTemplateBuilderComponent implements OnInit, OnChanges {
     permissionsUrl = '';
     historyUrl = '';
 
+    enableNewBuilder$: Observable<{ enableNewBuilder: boolean }>;
+
+    constructor(private readonly dotPropertiesService: DotPropertiesService) {}
     ngOnInit() {
+        this.enableNewBuilder$ = this.dotPropertiesService
+            .getKey(FeaturedFlags.FEATURE_FLAG_TEMPLATE_BUILDER)
+            .pipe(map((result) => ({ enableNewBuilder: result && result === 'true' })));
+
         this.permissionsUrl = `/html/templates/permissions.jsp?templateId=${this.item.identifier}&popup=true`;
         this.historyUrl = `/html/templates/push_history.jsp?templateId=${this.item.identifier}&popup=true`;
     }
