+++ core-web/apps/dotcms-ui/src/app/view/components/dot-edit-layout-designer/dot-edit-layout-designer.component.html
@@ -1,85 +1,70 @@
-<ng-container *ngIf="{ enableNewBuilder: enableNewBuilder$ | async } as context">
-    <ng-container *ngIf="context.enableNewBuilder; else oldTemplate">
-        <h1 data-testId="new-template-builder">New Template Builder</h1>
-    </ng-container>
+<dot-theme-selector
+    *ngIf="themeDialogVisibility"
+    [value]="currentTheme"
+    (selected)="changeThemeHandler($event)"
+    (shutdown)="closeThemeDialog()"
+>
+</dot-theme-selector>
+<dot-secondary-toolbar>
+    <div class="main-toolbar-left">
+        <dot-edit-page-info
+            class="flex align-items-center gap-2"
+            [apiLink]="apiLink"
+            [title]="title"
+            [url]="url"
+        ></dot-edit-page-info>
+    </div>
 
-    <ng-template #oldTemplate>
-        <dot-theme-selector
-            *ngIf="themeDialogVisibility"
-            [value]="currentTheme"
-            (selected)="changeThemeHandler($event)"
-            (shutdown)="closeThemeDialog()"
+    <div class="main-toolbar-right flex align-items-center gap-3">
+        <dot-global-message></dot-global-message>
+        <button
+            class="dot-edit-layout__toolbar-action-themes p-button-text p-button-vertical"
+            *ngIf="!currentTheme; else noPermission"
+            [label]="'editpage.layout.theme.button.label' | dm"
+            [pTooltip]="'org.dotcms.frontend.content.submission.not.proper.permissions' | dm"
+            (click)="themeDialogVisibility = !themeDialogVisibility"
+            disabled="false"
+            pButton
+            tooltipPosition="left"
+            type="button"
         >
-        </dot-theme-selector>
-        <dot-secondary-toolbar>
-            <div class="main-toolbar-left">
-                <dot-edit-page-info
-                    class="flex align-items-center gap-2"
-                    [apiLink]="apiLink"
-                    [title]="title"
-                    [url]="url"
-                ></dot-edit-page-info>
-            </div>
-
-            <div class="main-toolbar-right flex align-items-center gap-3">
-                <dot-global-message></dot-global-message>
-                <button
-                    class="dot-edit-layout__toolbar-action-themes p-button-text p-button-vertical"
-                    *ngIf="!currentTheme; else noPermission"
-                    [label]="'editpage.layout.theme.button.label' | dm"
-                    [pTooltip]="
-                        'org.dotcms.frontend.content.submission.not.proper.permissions' | dm
-                    "
-                    (click)="themeDialogVisibility = !themeDialogVisibility"
-                    disabled="false"
-                    pButton
-                    tooltipPosition="left"
-                    type="button"
-                >
-                    <span class="p-button-icon material-icons">brush</span>
-                </button>
-                <ng-template #noPermission>
-                    <button
-                        class="dot-edit-layout__toolbar-action-themes p-button-text p-button-vertical"
-                        [label]="currentTheme?.name || 'Themes'"
-                        (click)="themeDialogVisibility = !themeDialogVisibility"
-                        pButton
-                        type="button"
-                    >
-                        <span class="p-button-icon material-icons">brush</span>
-                    </button>
-                </ng-template>
-                <dot-layout-properties [group]="form.get('layout')"></dot-layout-properties>
-                <dot-icon-button
-                    [size]="40"
-                    (click)="addGridBox()"
-                    float
-                    icon="add"
-                ></dot-icon-button>
-                <span class="toolbar__sep" *ngIf="!apiLink || saveAsTemplate"></span>
-                <button
-                    class="p-button-secondary"
-                    *ngIf="!apiLink"
-                    [disabled]="disablePublish"
-                    [label]="'Publish' | dm"
-                    (click)="onSaveAndPublish()"
-                    data-testid="publishBtn"
-                    pButton
-                    type="button"
-                ></button>
-                <input
-                    class="dot-edit-layout__toolbar-template-name"
-                    #templateName
-                    *ngIf="saveAsTemplate"
-                    formControlName="title"
-                    pInputText
-                    placeholder="{{ 'editpage.layout.toolbar.template.name' | dm }} *"
-                    type="text"
-                />
-            </div>
-        </dot-secondary-toolbar>
-        <form [formGroup]="form" (ngSubmit)="onSave()" novalidate>
-            <dot-layout-designer formGroupName="layout"></dot-layout-designer>
-        </form>
-    </ng-template>
-</ng-container>
+            <span class="p-button-icon material-icons">brush</span>
+        </button>
+        <ng-template #noPermission>
+            <button
+                class="dot-edit-layout__toolbar-action-themes p-button-text p-button-vertical"
+                [label]="currentTheme?.name || 'Themes'"
+                (click)="themeDialogVisibility = !themeDialogVisibility"
+                pButton
+                type="button"
+            >
+                <span class="p-button-icon material-icons">brush</span>
+            </button>
+        </ng-template>
+        <dot-layout-properties [group]="form.get('layout')"></dot-layout-properties>
+        <dot-icon-button [size]="40" (click)="addGridBox()" float icon="add"></dot-icon-button>
+        <span class="toolbar__sep" *ngIf="!apiLink || saveAsTemplate"></span>
+        <button
+            class="p-button-secondary"
+            *ngIf="!apiLink"
+            [disabled]="disablePublish"
+            [label]="'Publish' | dm"
+            (click)="onSaveAndPublish()"
+            data-testid="publishBtn"
+            pButton
+            type="button"
+        ></button>
+        <input
+            class="dot-edit-layout__toolbar-template-name"
+            #templateName
+            *ngIf="saveAsTemplate"
+            formControlName="title"
+            pInputText
+            placeholder="{{ 'editpage.layout.toolbar.template.name' | dm }} *"
+            type="text"
+        />
+    </div>
+</dot-secondary-toolbar>
+<form [formGroup]="form" (ngSubmit)="onSave()" novalidate>
+    <dot-layout-designer formGroupName="layout"></dot-layout-designer>
+</form>
