+++ core-web/apps/dotcms-ui/src/app/view/components/dot-edit-layout-designer/dot-edit-layout-designer.component.spec.ts
@@ -25,12 +25,7 @@ import { DotEditLayoutService } from '@dotcms/app/api/services/dot-edit-layout/d
 import { DotHttpErrorManagerService } from '@dotcms/app/api/services/dot-http-error-manager/dot-http-error-manager.service';
 import { DotRouterService } from '@dotcms/app/api/services/dot-router/dot-router.service';
 import { DotTemplateContainersCacheService } from '@dotcms/app/api/services/dot-template-containers-cache/dot-template-containers-cache.service';
-import {
-    DotEventsService,
-    DotMessageService,
-    DotPropertiesService,
-    DotThemesService
-} from '@dotcms/data-access';
+import { DotEventsService, DotMessageService, DotThemesService } from '@dotcms/data-access';
 import { DotTheme } from '@dotcms/dotcms-models';
 import {
     cleanUpDialog,
@@ -95,7 +90,6 @@ let component: DotEditLayoutDesignerComponent;
 let fixture: ComponentFixture<DotEditLayoutDesignerComponent>;
 let dotThemesService: DotThemesService;
 let dotEditLayoutService: DotEditLayoutService;
-let dotPropertiesService: DotPropertiesService;
 
 describe('DotEditLayoutDesignerComponent', () => {
     beforeEach(() => {
@@ -158,12 +152,6 @@ describe('DotEditLayoutDesignerComponent', () => {
                         loading: jasmine.createSpy(),
                         customDisplay: jasmine.createSpy()
                     }
-                },
-                {
-                    provide: DotPropertiesService,
-                    useValue: {
-                        getKey: () => of('false')
-                    }
                 }
             ]
         });
@@ -172,7 +160,6 @@ describe('DotEditLayoutDesignerComponent', () => {
         component = fixture.componentInstance;
         dotThemesService = TestBed.inject(DotThemesService);
         dotEditLayoutService = TestBed.inject(DotEditLayoutService);
-        dotPropertiesService = TestBed.inject(DotPropertiesService);
     });
 
     describe('edit layout', () => {
@@ -410,23 +397,6 @@ describe('DotEditLayoutDesignerComponent', () => {
                     width: ''
                 }
             });
-        });
-    });
-
-    describe('New Template Builder', () => {
-        beforeEach(() => {
-            component.layout = mockDotLayout();
-            spyOn(dotPropertiesService, 'getKey').and.returnValue(of('true'));
-            fixture.detectChanges();
-        });
-
-        it('should show new template builder component', () => {
-            fixture.detectChanges();
-            const component: DebugElement = fixture.debugElement.query(
-                By.css('[data-testId="new-template-builder"]')
-            );
-
-            expect(component).toBeTruthy();
         });
     });
