+++ core-web/libs/dotcms-models/src/lib/shared-models.ts
@@ -16,7 +16,8 @@ export enum ComponentStatus {
 
 export const enum FeaturedFlags {
     LOAD_FRONTEND_EXPERIMENTS = 'FEATURE_FLAG_EXPERIMENTS',
-    DOTFAVORITEPAGE_FEATURE_ENABLE = 'DOTFAVORITEPAGE_FEATURE_ENABLE'
+    DOTFAVORITEPAGE_FEATURE_ENABLE = 'DOTFAVORITEPAGE_FEATURE_ENABLE',
+    FEATURE_FLAG_TEMPLATE_BUILDER = 'FEATURE_FLAG_TEMPLATE_BUILDER_2'
 }
 
 export type DotDropdownSelectOption<T> = {
